requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
torch
sentence-transformers
faiss-cpu
